endDadosAnimacion roto despues del primer ataque

+++ heuristicas/heuristica.h
@@ -24,7 +24,7 @@ public:
 
     virtual float factorReagrupe(TEG::TurnoAtaques *turno,TEG::AccionReagrupe *acc);
     virtual float cantidadReagrupe(TEG::TurnoAtaques *turno,TEG::AccionReagrupe *acc);
-    virtual float factorAtaque(TEG::TurnoAtaques *turno,TEG::AccionAtaque *acc);
+    virtual int factorAtaque(TEG::TurnoAtaques *turno,TEG::AccionAtaque *acc);
     virtual float cantidadAtaque(TEG::TurnoAtaques *turno,TEG::AccionAtaque *acc);
     virtual int factorFichas(TEG::TurnoFichas *turno, TEG::Pais *pais);
